Uses species names for the most common species in Ensembl queries
So far, the taxonomy IDs were used for both input and output species. However, today it started
to fail when using the taxonomy ID for Drosophila as input species (although it was working a few
days ago). Thus, for the most common species (those in list-species) the species names are used
for doing the queries.